+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Improve resource card layout [#65](https://github.com/IN-CORE/incore-studio/issues/65)
 - Disable unused buttons from navbar [#215](https://github.com/IN-CORE/incore-studio/issues/215)
+- Replace map preview with geo-explorer [#218](https://github.com/IN-CORE/incore-studio/issues/218)
 
 ## [Beta-1] - 05-01-2025